app/requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
authorizenet==1.1.3
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
decorator==4.4.2
Django==3.0.7
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
gunicorn==20.0.4
httplib2==0.18.1
idna==2.8
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
lxml==4.6.1
Markdown==3.2.2
MarkupSafe==1.1.1
multidict==4.7.6
oauthlib==3.1.0
openapi-codec==1.3.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
python-decouple==3.3
python-graphql-client==0.3.1
pytz==2020.1
PyXB==1.2.5
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
traitlets==5.0.4
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
websockets==8.1
yarl==1.6.2
